// apps/frontend/src/types/LibraryListResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FolderView } from "./FolderView";
import type { LibrarySkill } from "./LibrarySkill";
import type { TagView } from "./TagView";

export type LibraryListResponse = { libraryRevision: string | null, hasUncommittedChanges: boolean, pendingResolution: boolean, skills: Array<LibrarySkill>, folders: Array<FolderView>, tags: Array<TagView>, organizationRevision: string, };